+++ src/js/pyodide.js
@@ -170,7 +170,7 @@ function finalizeBootstrap(config) {
 
   Module.sys = Module.runPythonInternal("import sys; sys");
   Module.sys.setrecursionlimit(calculateRecursionLimit());
-  Module.sys.path.insert(0, config.homeDirectory);
+  Module.sys.path.insert(0, config.homedir);
 
   // Set up globals
   let globals = Module.runPythonInternal("import __main__; __main__.__dict__");
@@ -210,7 +210,7 @@ function finalizeBootstrap(config) {
  *
  * @param {string} config.indexURL - The URL from which Pyodide will load
  * packages
- * @param {string} config.homeDirectory - The home directory which Pyodide will use inside virtual file system
+ * @param {string} config.homedir - The home directory which Pyodide will use inside virtual file system
  * Default: /home/pyodide
  * @param {boolean} config.fullStdLib - Load the full Python standard library.
  * Setting this to false excludes following modules: distutils.
@@ -230,7 +230,7 @@ export async function loadPyodide(config) {
     fullStdLib: true,
     jsglobals: globalThis,
     stdin: globalThis.prompt ? globalThis.prompt : undefined,
-    homeDirectory: "/home/pyodide",
+    homedir: "/home/pyodide",
   };
   config = Object.assign(default_config, config);
   if (globalThis.__pyodide_module) {
@@ -257,7 +257,7 @@ export async function loadPyodide(config) {
   let packageIndexReady = initializePackageIndex(baseURL);
 
   setStandardStreams(config.stdin, config.stdout, config.stderr);
-  setHomeDirectory(config.homeDirectory);
+  setHomeDirectory(config.homedir);
 
   Module.locateFile = (path) => baseURL + path;
   let moduleLoaded = new Promise((r) => (Module.postRun = r));